requirements.txt
absl-py==2.0.0
aiofiles==23.1.0
aiogram==3.1.1
aiohttp==3.8.5
aioschedule==0.5.2
aiosignal==1.3.1
annotated-types==0.5.0
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.1.0
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
filelock==3.12.4
Flask==3.0.0
Flask-Cors==4.0.0
flatbuffers==23.5.26
frozenlist==1.4.0
fsspec==2023.9.2
gast==0.5.4
google-auth==2.23.2
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.59.0
h5py==3.9.0
huggingface-hub==0.16.4
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
keras==2.14.0
libclang==16.0.6
magic-filter==1.0.12
Markdown==3.4.4
MarkupSafe==2.1.3
ml-dtypes==0.2.0
mpmath==1.3.0
multidict==6.0.4
networkx==3.1
numpy==1.26.0
oauthlib==3.2.2
openai==0.28.1
opt-einsum==3.3.0
packaging==23.2
protobuf==4.24.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==2.3.0
pydantic_core==2.6.3
pyTelegramBotAPI==4.14.0
python-dotenv==1.0.0
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.3.3
six==1.16.0
sympy==1.12
tensorboard==2.14.1
tensorboard-data-server==0.7.1
tensorflow==2.14.0
tensorflow-estimator==2.14.0
tensorflow-intel==2.14.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
tokenizers==0.14.0
torch==2.0.1
tqdm==4.66.1
transformers==4.34.0
typing_extensions==4.7.1
urllib3==2.0.6
watchdog==3.0.0
Werkzeug==3.0.0
wrapt==1.14.1
yarl==1.9.2
